Stop ignoring argument type and invalid type error types.

+++ app_dart/lib/src/request_handlers/postsubmit_luci_subscription.dart
@@ -191,7 +191,7 @@ class PostsubmitLuciSubscription extends SubscriptionHandler {
       log.info('Updating check status for ${target.getTestName}');
       await githubChecksService.updateCheckStatus(
         build: build,
-        checkRunId: userDataMap['check_run_id'],
+        checkRunId: userDataMap['check_run_id'] as int,
         luciBuildService: scheduler.luciBuildService,
         slug: commit.slug,
       );
